minor refactor for geo attr crosswalk
